refactor: create send command helper
